+++ src/bloom_filter/bloom_filter_calculate_size.c
@@ -25,7 +25,7 @@
  * \return size in bytes the filter would need to be to meet the
  *         target error rate.
  */
-size_t bloom_filter_calculate_size(uint32_t num_expected_entries,
+size_t bloom_filter_calculate_size(size_t num_expected_entries,
     float target_error_rate)
 {
     size_t m_bytes;
